Update patch set 2
Patch Set 2:

> something broken with the build right now?

If you set BUILD_NUMBER to a constant, the ninja cache is always regenerated because it thinks CleanSpec changed.  This is because we have a symlinked dir in hw/qcom/audio-caf/apq8084:

legacy -> ../msm8974/legacy

That dir contains a CleanSpec.mk.  findleaves.py does not find it because of:

https://github.com/LineageOS/android_build/commit/7829ebfebd6dfd6aacfa691a2f7ccf7ce853acd6#diff-a73dbaaf22a923023901ec4fdc7eb666

However, ckati emulates findleaves.py when evaluating the ninja cache and it finds the symlinked CleanSpec.mk.

I just posted a review upstream to fix it:

https://android-review.googlesource.com/c/platform/build/kati/+/620686

Patch-set: 2