Add prebuilts for py3 embedded launcher
This triggers the distribution of the py3-launcher64 and
py3-launcher-autorun64 binaries, which are the launchers embedded within
python binaries using `embedded_launcher: true`. It also stores the
python3 stdlib under the common dist outputs so that the binaries and
stdlib can be updated in sync with each other.

These will be used in future CLs to stop building python3 for the host
on master, but to use prebuilts from prebuilts/build-tools.

Test: prebuilts/build-tools/build-prebuilts.sh
Change-Id: Ia00c1947d26e9c9ffaa94e37e47da6231b9baa16